+++ src/rc_be_/index.d.ts
@@ -1,4 +1,4 @@
-import type { Be, be_opts_T, Ctx } from '../be_'
+import type { Be, be_params_T, Ctx } from '../be_'
 import { be__val__T } from '../be_'
 export declare const be_m_set_key:unique symbol
 /**
@@ -13,7 +13,7 @@ export declare function rc_be_<
 >(key:string, val_:rc_be__val__T<Out>):rc_be__return_T<Out>
 export declare type rc_be__return_T<
 	Out extends NonNullable<unknown>
-> = (ctx:Ctx, opts?:rc_be_opts_T)=>RcBe_return_T<Out>
+> = (ctx:Ctx, opts?:rc_be_params_T)=>RcBe_return_T<Out>
 export {
 	rc_be_ as _rc_be,
 	rc_be_ as _rc_b,
@@ -26,12 +26,12 @@ export interface val_this_T {
 	on_destroy:val_this_on_destroy_T
 	onDestroy:val_this_on_destroy_T
 }
-export interface rc_be_opts_T extends be_opts_T {
+export interface rc_be_params_T extends be_params_T {
 	owner?:object
 }
 export declare type rc_be__val__T<
 	Out extends NonNullable<unknown>
-> = be__val__T<Out>&((this:val_this_T, ctx:Ctx, key:Be<Out>, opts?:rc_be_opts_T)=>Out)
+> = be__val__T<Out>&((this:val_this_T, ctx:Ctx, key:Be<Out>, opts?:rc_be_params_T)=>Out)
 export declare type rc_be_destroy_T = ()=>void
 export interface RcBe_return_T<Out extends unknown = unknown> {
 	value:Out
@@ -39,5 +39,5 @@ export interface RcBe_return_T<Out extends unknown = unknown> {
 }
 export declare type RcBe<
 	Out extends NonNullable<unknown>
-> = (ctx:Ctx, opts?:be_opts_T)=>RcBe_return_T<Out>
+> = (ctx:Ctx, opts?:be_params_T)=>RcBe_return_T<Out>
 export declare type RcB<Out extends NonNullable<unknown>> = RcBe<Out>